pre proyecto final parte 2

+++ src/service/cart.service.js
@@ -6,6 +6,7 @@ import {generateDocument,repetedDoc} from "../utlis/error/info.error.js"
 import logger from "../logger/factory.js"
 import UserClass from "../DAO/mongo/user.dao.js"
 import mongoose from "mongoose"
+import { sendMailActionOfAdmnin } from "./mail.service.js"
 
 const instanceUser = new UserClass()
 
@@ -13,7 +14,11 @@ export async function generateNewCart(email){
     const instanceUser =new UserClass()
     try{
         const getUser = await instanceUser.getUser({email,})
-        if(getUser.cart) return [`Ya tienes un carrito en tu cuenta ${getUser.firts_name}`,false,403]
+        if(getUser.cart){
+            const idCartString = getUser.cart._id.toString()
+            return [`Ya tienes un carrito en tu cuenta ${getUser.firts_name}, cuyo id es ${idCartString}`,false,403]
+        } 
+        if(getUser.rol == "admin") return ["Los admin no pueden crearse carritos",false,403]
        const cart =  await instanceCart.postCart()
         await instanceUser.setNewProperty(getUser._id,"cart",cart._id)
         return cart._id.toString()
@@ -32,14 +37,26 @@ export async function getCartById(cid){
 }
 
 export async function corroborateQuery(limit,sort,page,query){
-
-    !limit && (limit = 10)
-    !sort && (sort = - 1)
-    !page && (page = 1)
-    if(query && typeof query !== "object") return ["La query debe ser un objecto",false,403]
-
     try{
-        return await instanceCart.getCartsByQuery(limit,sort,page,query)
+        const totalCarts = await instanceCart.getCartsCount()
+        const numberPage = page? Number(page):" "
+        const numberLimit = limit?Number(limit):" "
+        const numberSort = sort? Number(sort):" "
+
+        if(!numberLimit || !numberPage || !numberSort) return ["Las querys limit,page,sort deben enviarse como numeros",false,403]   
+        
+        if(numberPage > totalCarts && numberLimit > totalCarts) return [`Las páginas o limites no pueden ser mayores a los carritos ${totalCarts} existentes`,false,403]
+        if((typeof numberSort == "number") && (numberSort !== -1 && numberSort !== 1)) return ["La query sort solo acepta 1(ascendente) o -1(descendente)",false,403]
+
+        numberLimit == " " && (limit = 10)
+        numberSort == " " && (sort = - 1)
+        numberPage == " " && (page = 1)
+
+        if(query && typeof query !== "object") return ["La query debe ser un objecto",false,403]
+
+        const querys = await instanceCart.getCartsByQuery(numberLimit,numberSort,numberPage,query)
+        if(querys.totalDocs == 0) return [`La ${query} no coincide con ningún criterio de búsqueda`,false,403]
+        return querys
     }catch(err){
         throw new Error(err)
     }
@@ -49,9 +66,10 @@ export async function postProduct(cid,pid,quankity,email){
     try{
         const user = await instanceUser.getUser({email,})
         if(user.rol == "premium"){
-            const creadorId = creator.id
-            const products = await productsService.getProducts({[creadorId]:user._id.toString()})
-            const filterProduct = products.find(item => item._id.toString() == pid)
+            const productsUser = user.products_created
+            const filterProduct = productsUser.find(item => item.toString() == pid)
+            /* const products = await productsService.getProducts({[creadorId]:user._id.toString()})
+            const filterProduct = products.find(item => item._id.toString() == pid) */
             if(filterProduct) return ["No puedes añadir productos tuyos",false,403]
         }
         const cartId = await instanceCart.getCartById(cid)
@@ -69,6 +87,9 @@ export async function postProduct(cid,pid,quankity,email){
         }
 
         if(!product.status) return ["El status del producto es falso",false,403]
+        const verProductoRepetido = cartId.products.findIndex(item => item.product._id.toString() == pid)
+        if(product.stock == 0) return ["No es posible añadir productos sin stock",false,403]
+        if(verProductoRepetido) return ["No podes añadir el mismo producto",false,403]
 
         const verCantidad = cartId.products.find(item => item.quankity == quankity && item.id == idMongoString)
         if(verCantidad){
@@ -80,7 +101,7 @@ export async function postProduct(cid,pid,quankity,email){
             })
             return ["No puedes añadir la misma cantidad",false,409]
         }
-        cartId.products.push({product:pid,quankity})
+        verProductoRepetido == -1 && cartId.products.push({product:pid,quankity})
         logger.info("cantidad actualizada","sss")
         return  await instanceCart.replazeCart(cartId._id,cartId)
     }catch(error){
@@ -89,8 +110,9 @@ export async function postProduct(cid,pid,quankity,email){
     }
 }
 
-export async function patchProducts(cid,id,cantidad){
+export async function patchProducts(cid,id,cantidad,rol){
     try{
+        const objIdCart =  mongoose.Types.ObjectId(cid)
         const cart = await instanceCart.getCartById(cid)
         const filter = cart.products.findIndex((item) => item.product._id == id)
         if(filter == -1 ||!cart){
@@ -107,7 +129,11 @@ export async function patchProducts(cid,id,cantidad){
         obj.quankity = cantidad
         cart.products.splice(filter,1,obj)
         const upDateProduct = await instanceCart.replazeCart(cart._id,cart)
-        
+        if(rol == "admin"){
+            const user = await instanceUser.getUser({cart:cid})
+            const text = `Hola ${user.firts_name} ${user.last_name} le informamos que nuestro equipo le ha borrado el producto con el id ${pid} de su carrito.Comuniquese con nuestro soporte en caso de haber sido una confusión`
+            sendMailActionOfAdmnin(user.email,text)
+        }
         return upDateProduct
     }catch(error){
         req.logger.error(error)
@@ -116,26 +142,43 @@ export async function patchProducts(cid,id,cantidad){
     
 }
 
-export async function deleteProductCart(cid,pid){
+export async function deleteProductCart(cid,pid,rol){
     const objectIdCart = mongoose.Types.ObjectId(cid)
-    const objectIdProduct = mongoose.Types.ObjectId(pid)
 
     if(pid.length !== 24) return [`El ${pid} no es un id válido.Recuerda que tiene que ser de 24 carácteres alfanuméricos`,false,403]
 
+    const objectIdProduct = mongoose.Types.ObjectId(pid)
     try{
-       return await instanceCart.deleteProduct(objectIdCart,objectIdProduct)
+        const cartById = await instanceCart.getCartById(cid)
+        const productExist = cartById.products.find(item => item.product._id.toString() == pid)
+
+        if(!productExist) return ["El producto no existe",false,404]
+
+        const deleteProduct = await instanceCart.deleteProduct(objectIdCart,objectIdProduct)
+        console.log(rol)
+        if(rol == "admin"){
+            const user = await instanceUser.getUser({cart:cid})
+            console.log(rol,user)
+            const text = `Hola ${user.firts_name} ${user.last_name} le informamos que nuestro equipo le ha borrado el producto con el id ${pid} de su carrito.Comuniquese con nuestro soporte en caso de haber sido una confusión`
+            sendMailActionOfAdmnin(user.email,text)
+        }
+        return deleteProduct
     }catch(err){
         throw new Error(err)
     }
 }
 
-export async function deleteCart(cid,email){
+export async function deleteCart(cid,email,rol){
     const objectIdCart = mongoose.Types.ObjectId(cid)
     try{
-        const user = await instanceUser.getUser({email,})
-
-         await instanceCart.deleteCart(objectIdCart)
+    const user = await instanceUser.getUser({cart:objectIdCart})
+    if(rol == "admin"){
+        const text = `Hola ${user.firts_name} ${user.last_name} le informamos que nuestro equipo le ha borrado el carrito.Comuniquese con nuestro soporte en caso de haber sido una confusión`
+        sendMailActionOfAdmnin(user.email,text)
+    }
+        await instanceCart.deleteCart(objectIdCart)
         return await instanceUser.unsetPropery(user._id,"cart")
+
     }catch(err){
         throw new Error(err)
     }
